Move checkpoint public API to RevenueCatUI
